requirements.txt
alabaster==0.7.13
Babel==2.13.1
blinker==1.7.0
boto3==1.33.11
botocore==1.33.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
cryptography==41.0.7
dnspython==2.4.2
docutils==0.20.1
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pymongo==4.6.0
pymongo-auth-aws==1.1.0
pymongocrypt==1.7.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
s3transfer==0.8.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tomli==2.0.1
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
